Normalizing all DTOs, Requests and Responses to be records - pt 1 | Changed product prices to be stored as decimals (no precision point given - default 2 digits?)

+++ GSW/GSW-Core/DTOs/Product/ProductDTO.cs
@@ -14,7 +14,7 @@ namespace GSW_Core.DTOs.Product
         string Name,
         string Description,
         DateTime ReleaseDate,
-        int Price,
+        decimal Price,
         ICollection<DeveloperDTO> Developers,
         ICollection<PublisherDTO> Publishers,
         ICollection<GenreDTO> Genres,
